same for verifier endpoints
