+++ services/showcase/api/endpoints/routes/sandbox.py
@@ -18,6 +18,9 @@ from api.endpoints.routes.job_applicant import router as applicants_router
 from api.endpoints.routes.line_of_business import router as lobs_router
 from api.endpoints.routes.line_of_business_issuer import router as lobs_issuer_router
 from api.endpoints.routes.line_of_business_holder import router as lobs_holder_router
+from api.endpoints.routes.line_of_business_verifier import (
+    router as lobs_verifier_router,
+)
 
 router = APIRouter()
 logger = logging.getLogger(__name__)
@@ -39,6 +42,11 @@ router.include_router(
 router.include_router(
     lobs_holder_router,
     tags=["line_of_business/holder"],
+    prefix="/sandboxes/{sandbox_id}/lobs/{lob_id}",
+)
+router.include_router(
+    lobs_verifier_router,
+    tags=["line_of_business/verifier"],
     prefix="/sandboxes/{sandbox_id}/lobs/{lob_id}",
 )
